Update nanomine converter
